+++ baguni-batch/src/test/java/baguni/batch/domain/link/service/LinkServiceTest.java
@@ -1,4 +1,6 @@
-package baguni.batch.domain.feed;
+package baguni.batch.domain.link.service;
+
+import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -7,24 +9,24 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 
 import baguni.BaguniBatchApplication;
-import baguni.batch.domain.feed.util.FeedApi;
-import baguni.batch.domain.feed.service.FeedService;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @ActiveProfiles("local")
 @SpringBootTest(classes = BaguniBatchApplication.class)
-public class RssRetrieveTest {
+class LinkServiceTest {
 
 	@Autowired
-	FeedService feedService;
+	LinkService linkService;
 
 	/**
 	 * DB에 블로그 목록이 반드시 존재해야 한다.
 	 */
 	@Test
-	@DisplayName("전체 블로그 feed를 잘 획득하는지 테스트")
+	@DisplayName("링크 분석 후 결과를 DB에 잘 저장하는지 테스트")
 	public void test() {
-		feedService.saveBlogArticleLinks();
+		var url = "https://techblog.lycorp.co.jp/ko/2024-frontend-global-workshop-recap";
+		linkService.analyzeAndUpdateLink(url);
 	}
+
 }
